refactor(view): split design_import.cpp JSON ser/de into design_ir_json.cpp
Relocation-only first cut of the design-import TU split (PR-1 of
planning/2026-05-29-frontend-ir-refactor-plan.md). Moves the DesignIR JSON
serialize/deserialize band (~1150 lines) out of the 2681-line
design_import.cpp into a new design_ir_json.cpp, dropping design_import.cpp to
1526 lines. Six IR-parse / diagnostic primitives shared across the new TU
boundary (parse_ir_node, parse_ir_tokens, make_import_diagnostic,
is_asset_reference_key, audio_widget_from_id, promote_interactive_frames) are
declared in the private design_import_internal.hpp; the moved band is otherwise
byte-identical (no logic changes). serialize_design_ir / parse_design_ir_json
remain public API. All 18 design-import test binaries pass locally.

Skill-Update: skip skill=import-design reason="relocation-only TU split; no behavior or contract change to the import pipeline"
