+++ core/view/CMakeLists.txt
@@ -301,6 +301,7 @@ target_sources(pulp-view-core PRIVATE
     src/relative_point.cpp
     src/design_export.cpp
     src/design_import.cpp
+    src/design_ir_json.cpp
     src/design_import_v0_tsx.cpp
     src/design_import_native_common.cpp
     src/design_cpp_codegen.cpp
